+++ src/Resume.Models/OnlineProfile.cs
@@ -3,9 +3,7 @@ namespace Resume.Models
     public class OnlineProfile
     {
         public string Network { get; set; }
-
         public string Username { get; set; }
-
         public string Url { get; set; }
     }
 }