+++ src/Resume.Models/Skill.cs
@@ -5,9 +5,7 @@ namespace Resume.Models
     public class Skill
     {
         public string Name { get; set; }
-
         public string Level { get; set; }
-
         public List<string> Keywords { get; set; }
     }
 }