+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [v0.3.0] - 2019-01-19
+### Fixed
+- Page title no longer hard coded
+
+### Changed
+- Use a page model for resume templates
+
 ## [v0.2.0] - 2019-01-17
 ### Changed
 - Set default A4 and margins of 72 points for pdf converter